2554. Maximum Number of Integers to Choose From a Range I.
